Use float instead of double for camera position.

+++ src/engine/parse/xml/camera/projection.cpp
@@ -6,12 +6,12 @@
 namespace engine::parse::xml {
 
 auto parse_projection(rapidxml::xml_node<> const* const node) noexcept
-    -> cpp::result<glm::dvec3, parse_err>
+    -> cpp::result<glm::vec3, parse_err>
 {
     return glm::vec3 {
-        TRY_RESULT(parse_number_attr<double>(node, "fov")),
-        TRY_RESULT(parse_number_attr<double>(node, "near")),
-        TRY_RESULT(parse_number_attr<double>(node, "far")),
+        TRY_RESULT(parse_number_attr<float>(node, "fov")),
+        TRY_RESULT(parse_number_attr<float>(node, "near")),
+        TRY_RESULT(parse_number_attr<float>(node, "far")),
     };
 }
